refactor: 파일 경로 수정 및 type import 선언
