+++ just-plan/src/app/(needLogin)/_components/DetailPlace/DetailPlace.tsx
