+++ just-plan/src/components/AddedPlaceCard/AddedPlaceCard.types.ts
@@ -1,13 +1,5 @@
-export interface IPlan {
-  id: number;
-  date: string;
-  image: string;
-  title: string;
-  category: string;
-  address: string;
-  time: string | null;
-}
+import { ILocationInfo } from "@/types/plan.types";
 
 export interface IProps {
-  item: IPlan;
+  item: ILocationInfo;
 }
