+++ just-plan/src/components/DayPlanCard/DayPlanCard.tsx
@@ -1,6 +1,6 @@
 import React, { PropsWithChildren } from "react";
 import { Switch } from "../ui/switch";
-import { IProps } from "./DayPlanCard.types";
+import type { IProps } from "./DayPlanCard.types";
 import { Plan } from "@/mocks";
 import { AddedPlaceCard } from "..";
